+++ packages/core/src/errors.ts
@@ -98,7 +98,16 @@ export class ChainMismatchError extends Error {
 
 export class ChainNotConfiguredError extends Error {
   name = 'ChainNotConfigured'
-  message = 'Chain not configured'
+
+  constructor({
+    chainId,
+    connectorId,
+  }: {
+    chainId: number
+    connectorId: string
+  }) {
+    super(`Chain "${chainId}" not configured for connector "${connectorId}".`)
+  }
 }
 
 export class ConnectorAlreadyConnectedError extends Error {